HP Bar and Knockback
Implemented knockback functionality when the user hits the slime. Knockback force can be changed, but as of right now it looks and feels pretty decent.

I also added an HP bar that shows above the slime's icon at all times, and when the slime is hit, the bar adjusts based on the slime's new health, with a cooldown to show the bar being lowered after being damaged.